Exploration Testing Module/pyproject.toml
[project]
name = "exploration-testing-module"
version = "0.1.0"
description = ""
authors = [
    {name = "Raphael Mazzoleni"}
]
readme = "README.md"
requires-python = ">=3.8,<4.0"
dependencies = [
    "python-docx (>=1.1.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "ollama (>=0.4.8,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
